se refactorizo el codigo para que estos esten mas legibles y facil de entender
