+++ lib/views/storefonts_tab.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class StorefrontsTab extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('Storefronts'),
+    );
+  }
+}
