+++ lib/views/wishlist_tab.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class WishlistTab extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('Wishlist'),
+    );
+  }
+}
